🚧 very strange constant errors
