--- plugins/iconify.ts
@@ -1,5 +1,0 @@
-import { Icon } from '@iconify/vue'
-import { defineNuxtPlugin } from '#app'
-export default defineNuxtPlugin((nuxtApp) => {
-  nuxtApp.vueApp.component('Icon', Icon)
-})
